Create WATERCOOLER2.py
Problem statement: https://www.codechef.com/problems/WATERCOOLER2

Cost of using a cooler per month is given and the cost of purchasing it is also given. Print the number of months the cooler can be used so that the cost of using it is less than the cost of purchasing the cooler.