Create CYBV.py
Problem statement: https://www.codechef.com/problems/CYBV

Given n number of kids and k number of weapons. Find the minimum number of weapons that can be distribute between the kids so that the difference between the kid with the most number of weapons and least number of weapons is never greater than 1.