Create DIET.py
Problem statement: https://www.codechef.com/problems/DIET?tab=statement

Given the amount of protein bought on a certain date and the amount of protein consumed on a certain day; print if the amount of protein will be able to meet the requirement of the person. If not, print the day on which the person will not be able to meet his protein requirement.