Create AVG.py
Problem statement: https://www.codechef.com/problems/AVG?tab=statement

Given an array of size N. K elements have been deleted from this array. The arithmetic average of the array before deleting was V. All the deleted elements were of the same value. Find out if the value of the deleted element or output -1 if the given V is impossible to obtain given the other values.