Create HIGHFREQ.py
Problem statement: https://www.codechef.com/LP2TO308/problems/HIGHFREQ

Given an array with elements in it at a certain frequency. We are allowed to change a subsequence of this array, where all the elements of the subsequence are the same. Find the minimum possible frequency of the most frequent element of the array, after the above described operation is performed.