Create WALK.py
Problem statement: https://www.codechef.com/problems/WALK?tab=statement

Given an array of integers, find the minimal possible integer that when reduced by 1 for every given element of the array, is never less than any of the elements of the array.

Status: This problem wasn't solved by viewing the solution. But, it was first attempted in PHP that was giving a runtime error due to overflow caused by addition of large integers. But when it was implemented in Python, that error didn't occur probably due to no overflow.