Creating CHEFSUM.py
Problem statement: https://www.codechef.com/problems/CHEFSUM?tab=statement

Problem description: Need to find the index at which the sum of the
prefix sum and suffix sum of a given array is minimum.

solution status: It was solved without looking at the solution. The main
issue was with removing a O(n**2) complexity to O(n).