Create STRONGTABLE.py
Problem description: https://www.codechef.com/problems/STRONGTABLE?tab=statement

Find the maximum weight a table can withstand by using some non-empty
subset of table legs.

Status: Had to read the hints and look at the solution to understand
that the table legs had to be sorted in descending order. Once that's
done we can multiply the weight that a leg can bear with it's position.