Create two-sum.py
This is my accepted solution to the two-sum problem in Leetcode: https://leetcode.com/problems/two-sum/

A nested for loop was used for it. The outer for loop goes till the end of the input array. The inner for loop goes 1 less than the value of the outer for loop. This way the loops never go out of bound. 

Then the if statement adds two elements in the loop together to check if they add up to be equal to the target. If they do, then they are appended to the output array. After which we break from the loop. 

There's a break at the level of the inner for loop to exit the outer for loop. 

Since there's only one right answer in every input array, the array is outputted.