Create adding-spaces-to-a-string.py
Solution to the following Leetcode problem: https://leetcode.com/problems/adding-spaces-to-a-string/

Solution description:
Use a variable to store parts of the string. These parts are stored in the list. The start value changes based on the contents of the spaces variable. The last part of the string slice is obtained after we have run through the spaces list. After that we can join the contents of the list by using the map function inside the join function and returning that value.